+++ CHANGELOG.md
@@ -7,9 +7,10 @@ Changes to the plugin.
 ### Fixed
 
 - Variables in multi-parentheses expressions like `*if (var1) or (var2)` are now indexed properly.
-- Variable references in choice options like `  #Option ${var}` are now indexed properly.
+- Variable references in choice options like `#Option ${var}` are now indexed properly.
 - Variable references in `*bug` commands are now indexed correctly.
 - Syntax highlighting now applied to multireplaces in string literals.
+- Syntax highlighting now applied to variables and multireplaces in bold and italicized sections.
 
 ## [1.4.0] - 2020-05-11